% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hitter_profile.R
\name{calculate_hitter_profile}
\alias{calculate_hitter_profile}
\title{Calculate stats of hitter over specified time}
\usage{
calculate_hitter_profile(
  firstName,
  lastName,
  startDate,
  endDate,
  playerIndex = 1
)
}
\arguments{
\item{firstName}{A string}

\item{lastName}{A string}

\item{startDate}{A string formatted YYYY-MM-DD}

\item{endDate}{A string formatted YYYY-MM-DD}

\item{playerIndex}{A positive integer}
}
\value{
The counting stats of a hitter over a specified period of time
}
\description{
Calculate stats of hitter over specified time
}
\examples{
calculate_hitter_profile("Juan","Soto", "2024-05-19", "2025-06-05", playerIndex = 6)
}
