% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spray_chart.R
\name{plot_spray_chart}
\alias{plot_spray_chart}
\title{Plotting spray chart over specified time}
\usage{
plot_spray_chart(firstName, lastName, startDate, endDate, playerIndex = 1)
}
\arguments{
\item{firstName}{A string}

\item{lastName}{A string}

\item{startDate}{A string formatted YYYY-MM-DD}

\item{endDate}{A string formatted YYYY-MM-DD}

\item{playerIndex}{A positive integer}
}
\value{
A table with all players that match the input name, and a spray chart plot of the specified time period data
}
\description{
Plotting spray chart over specified time
}
\examples{
plot_spray_chart("Juan","Soto", "2024-05-19", "2025-06-05", playerIndex = 6)
}
