% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zone_heatmap.R
\name{plot_zone_heatmap}
\alias{plot_zone_heatmap}
\title{Plotting zone heatmap of batting average over specified time}
\usage{
plot_zone_heatmap(firstName, lastName, startDate, endDate, playerIndex = 1)
}
\arguments{
\item{firstName}{A string}

\item{lastName}{A string}

\item{startDate}{A string formatted YYYY-MM-DD}

\item{endDate}{A string formatted YYYY-MM-DD}

\item{playerIndex}{A positive integer}
}
\value{
A table with all players that match the input name, and a heatmap plot of the date-range data
}
\description{
Plotting zone heatmap of batting average over specified time
}
\examples{
plot_zone_heatmap("Juan","Soto", "2024-05-19", "2025-06-05", playerIndex = 6)
}
